requirements.txt
streamlit>=1.28.0
yfinance>=0.2.18
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
feedparser>=6.0.10
python-dateutil>=2.8.2